requirements.txt
beautifulsoup4>=4.12.3
feedparser>=6.0.11
httpx>=0.27.0
lxml>=5.2.1
python-dotenv>=1.0.1
requests>=2.31.0
schedule>=1.2.1
